:sparkles: Feat: Profile Left Menu
