+++ src/pages/RecruitmentEntryPage.jsx
@@ -2,6 +2,7 @@ import React, { useState } from 'react';
 import Margin from '../components/Margin';
 import DynamicColorButton from '../components/DynamicColorButton';
 import DatePicker from 'react-datepicker';
+import { Link ,useNavigate} from "react-router-dom";
 import 'react-datepicker/dist/react-datepicker.css';
 import EditorComponent from '../components/ui/EditorComponent';
 import InputField from '../components/InputField';
@@ -22,7 +23,7 @@ function RecruitmentPostEntryPage() {
     maxAttendance: '',
 
   });
-
+  const navigate = useNavigate();
   const handleContentChange = (newContent) => {
     setFormData({ ...formData, content: newContent });
     // You can perform additional actions with the new content if needed
@@ -32,6 +33,7 @@ function RecruitmentPostEntryPage() {
     setFormData({ ...formData, [field]: e.target.value });
     console.log(formData);
   };
+  
 
 
   const regions = [
@@ -95,8 +97,12 @@ function RecruitmentPostEntryPage() {
     });
   };
 
+
+ 
+
   const handleBack = () => {
-    console.log('Register clicked. Form data:', formData);
+
+    navigate('/recruits')
   };
 
   return (
@@ -183,6 +189,7 @@ function RecruitmentPostEntryPage() {
               <DynamicColorButton color="blue" text="작성" btnstyle="" onClick={handleSave} />
             </div>
           </div>
+        
         </div>
       </div>
     </div>
